Revert "fix: param settings FAB obstructs StoryContent (#68)"
This reverts commit 8a3a3103480ce26a0d2fec1818c24b7c0c110d5d.
