+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.9.2"
+agp = "8.7.1"
 android-compileSdk = "35"
 android-minSdk = "26"
 android-targetSdk = "35"
