+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,7 @@
-#Fri May 02 19:34:04 ICT 2025
+#Tue Oct 29 23:06:22 CST 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.11.1-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.9-bin.zip
 networkTimeout=10000
 validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
