+++ modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/compose/CompositionLocal.kt
@@ -1,9 +1,5 @@
 package org.jetbrains.compose.storytale.gallery.compose
 
-import androidx.compose.runtime.staticCompositionLocalOf
-
 inline fun <reified T> noCompositionLocalProvided(): T {
     error("CompositionLocal ${T::class.simpleName} not present")
 }
-
-val LocalIsEmbeddedView = staticCompositionLocalOf { false }
