+++ modules/gallery/src/commonMain/kotlin/org/jetbrains/compose/storytale/gallery/material3/FullStoryView.kt
@@ -17,7 +17,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.DrawerState
 import androidx.compose.material3.DrawerValue
@@ -33,11 +32,9 @@ import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -49,7 +46,6 @@ import androidx.navigation.toRoute
 import androidx.window.core.layout.WindowWidthSizeClass
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.storytale.Story
-import org.jetbrains.compose.storytale.gallery.compose.LocalIsEmbeddedView
 import org.jetbrains.compose.storytale.gallery.ui.theme.UseCustomDensity
 import org.jetbrains.compose.storytale.storiesStorage
 
@@ -95,20 +91,14 @@ fun FullStorytaleGallery(
                 popExitTransition = { ExitTransition.None },
             ) {
                 composable<StoryScreen> {
-                    val showOverlayParameters = rememberSaveable { mutableStateOf(false) }
                     Column(modifier = Modifier.fillMaxSize()) {
                         GalleryTopAppBar(
                             appState = appState,
                             drawerState = drawerState,
-                            activeStory = activeStoryItem.value?.story,
-                            showOverlayParameters = showOverlayParameters,
+                            activeStoryName = activeStoryItem.value?.story?.name,
                         )
                         HorizontalDivider()
-                        StoryContent(
-                            activeStory = activeStoryItem.value?.story,
-                            showOverlayParameters = showOverlayParameters,
-                            modifier = Modifier.fillMaxSize(),
-                        )
+                        StoryContent(activeStoryItem.value?.story, modifier = Modifier.fillMaxSize())
                     }
                 }
             }
@@ -187,15 +177,12 @@ private fun DrawerContent(
     )
 }
 
-@Suppress("ktlint:compose:mutable-state-param-check")
 @Composable
 private fun GalleryTopAppBar(
     drawerState: DrawerState,
     appState: StorytaleGalleryAppState,
-    activeStory: Story?,
-    showOverlayParameters: MutableState<Boolean>,
+    activeStoryName: String?,
 ) {
-    val activeStoryName = activeStory?.name.orEmpty()
     val coroutineScope = rememberCoroutineScope()
     val currentWindowWidthClass = currentWindowAdaptiveInfo().windowSizeClass.windowWidthSizeClass
     val isExpanded = currentWindowWidthClass == WindowWidthSizeClass.EXPANDED
@@ -203,7 +190,7 @@ private fun GalleryTopAppBar(
     CenterAlignedTopAppBar(
         title = {
             AnimatedContent(activeStoryName) { title ->
-                Text(title)
+                Text(title ?: "")
             }
         },
         colors = TopAppBarDefaults.topAppBarColors(containerColor = MaterialTheme.colorScheme.surfaceContainerLowest),
@@ -239,36 +226,16 @@ private fun GalleryTopAppBar(
             }
         },
         actions = {
-            val widthClass = currentWindowAdaptiveInfo().windowSizeClass.windowWidthSizeClass
-
-            val isSmallWidth = widthClass == WindowWidthSizeClass.COMPACT
-            val useEmbeddedView = LocalIsEmbeddedView.current
-
             ThemeSwitcherIconButton(appState)
-            AnimatedVisibility(
-                (isSmallWidth || useEmbeddedView) && activeStory?.parameters?.isNotEmpty() == true,
-                enter = fadeIn(),
-                exit = fadeOut(),
-            ) {
-                IconButton(
-                    onClick = {
-                        showOverlayParameters.value = true
-                    },
-                ) {
-                    Icon(imageVector = Icons.Default.Settings, null)
-                }
-            }
         },
     )
 }
 
 @Composable
 internal fun ThemeSwitcherIconButton(appState: StorytaleGalleryAppState) {
-    IconButton(
-        onClick = {
-            appState.switchTheme(!appState.isDarkTheme())
-        },
-    ) {
+    IconButton(onClick = {
+        appState.switchTheme(!appState.isDarkTheme())
+    }) {
         AnimatedContent(targetState = appState.isDarkTheme()) { isDarkTheme ->
             if (isDarkTheme) {
                 Icon(imageVector = Light_mode, contentDescription = null)
